backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
scikit-learn>=1.3.2
numpy>=1.26.2
Pillow==10.2.0
reportlab==4.0.7
openpyxl==3.1.2
httpx==0.25.2
icrawler==0.6.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.0
joblib==1.3.2
